add: PinImageのWidget化
